requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
fastapi==0.108.0
Flask==3.0.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.3
pydantic_core==2.14.6
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
Werkzeug==3.0.1